deps/file-updater: Add file-updater helper library
This library allows plugins to update/sync data files from a remote
server.  Useful for updating services on the fly, or updating game
capture exceptions via a text file (which are the two planned uses I
have for this particular library).

Checks for a "package.json" file from a remote, with the value of "url"
and an array of files with their respective versions.  First, if any of
the local versions are newer than the cached version, it will replace
each older cached file with the local file that's newer.  After that, if
the remote version is newer than the cached version, it will download
those files to the cache and replace them.